pyproject.toml
[tool.poetry]
name = "robotic-researcher"
version = "0.1.0"
description = ""
authors = ["Felipe Rodríguez <farodrig92@gmail.com>"]
readme = "README.md"
packages = [{include = "robotic_researcher"}]

[tool.poetry.dependencies]
python = "^3.11"
rpaframework = "23.0.0"
typer = {extras = ["all"], version = "^0.9.0"}
yarl = "^1.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
ruff = "^0.0.269"

[tool.ruff]
# Decrease the maximum line length to 79 characters.
line-length = 100
src = ["src"]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "UP",  # pyupgrade
  "C90", # mccabe
  "I", # isort
  "N", # pep8-naming
  "ERA", # erradicate
  "PL", # pylint
  "TRY", # tryceratops 
  "FLY", # flynt 
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [
    "--import-mode=importlib",
]
